+++ database/install/myems_system_db.sql
@@ -831,22 +831,6 @@ CREATE TABLE IF NOT EXISTS `myems_system_db`.`tbl_microgrids_users` (
 CREATE INDEX `tbl_microgrids_users_index_1` ON  `myems_system_db`.`tbl_microgrids_users` (`microgrid_id`);
 
 -- ---------------------------------------------------------------------------------------------------------------------
--- Table `myems_system_db`.`tbl_microgrids_windturbines`
--- ---------------------------------------------------------------------------------------------------------------------
-DROP TABLE IF EXISTS `myems_system_db`.`tbl_microgrids_windturbines` ;
-
-CREATE TABLE IF NOT EXISTS `myems_system_db`.`tbl_microgrids_windturbines` (
-  `id` BIGINT NOT NULL AUTO_INCREMENT,
-  `name` VARCHAR(255) NOT NULL,
-  `uuid` CHAR(36) NOT NULL,
-  `microgrid_id` BIGINT NOT NULL,
-  `power_point_id` BIGINT NOT NULL,
-  `meter_id` BIGINT NOT NULL,
-  `capacity` DECIMAL(18, 3) NOT NULL,
-  PRIMARY KEY (`id`));
-CREATE INDEX `tbl_microgrids_windturbines_index_1` ON  `myems_system_db`.`tbl_microgrids_windturbines` (`name`);
-
--- ---------------------------------------------------------------------------------------------------------------------
 -- Table `myems_system_db`.`tbl_offline_meters`
 -- ---------------------------------------------------------------------------------------------------------------------
 DROP TABLE IF EXISTS `myems_system_db`.`tbl_offline_meters` ;